LayoutTests/platform/chromium/accessibility/scroll-to-make-visible-div-overflow-expected.txt

Tests that scrolling to make an element visible successfully scrolls an arbitrary HTML element that has CSS overflow set to 'scroll'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Upper Target
5000-pixel box
Lower Target
PASS container.scrollTop is 0
PASS container.scrollTop is >= 4952
PASS 5030 is >= container.scrollTop
PASS container.scrollTop is >= -76
PASS 2 is >= container.scrollTop
PASS successfullyParsed is true

TEST COMPLETE
